% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frgtracker.R
\name{calculate_final_grade}
\alias{calculate_final_grade}
\title{Calculate final grades for all students in a course}
\usage{
calculate_final_grade(courses, grades, course_ids)
}
\arguments{
\item{courses}{A dataframe containing component weights for each course}

\item{grades}{A dataframe containing grades for students}

\item{course_ids}{A vector of strings representing the course IDs for which
final grades should be calculated.}
}
\value{
A dataframe containing final grades for all students in a course.
}
\description{
For each student, the mark for each component will be multiplied with its
associated weight, and summed up altogether.
}
\examples{
calculate_final_grade(course_id = c("511"))
}
