% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frgtracker.R
\name{rank_courses}
\alias{rank_courses}
\title{Calculate students' course grades to rank courses in ascending/descending order by a
specified method.}
\usage{
rank_courses(
  courses,
  grades,
  method = c("mean", "1st-quantile", "median", "3rd-quantile"),
  descending = TRUE
)
}
\arguments{
\item{courses}{A dataframe containing component weights for each course}

\item{grades}{A dataframe containing grades for students}

\item{method}{one of "method", "median", "lst-quantile", "3rd-quantile", defining
the method for calculating the course rankings.}

\item{descending}{A logical value to decide if the rank should be in descending or
ascending order. Default to True}
}
\value{
A dataframe containing the rank for specified courses
}
\description{
Calculate students' course grades to rank courses in ascending/descending order by a
specified method.
}
\examples{
grades <- data.frame(
course_id = c("511"),
student_id = c("tom"),
lab1 = c(100),
lab2 = c(80)
)
courses <- data.frame(
course_id = c("511"),
lab1 = c(0.45),
lab2 = c(0.55)
)
rank_courses(courses, grades, "mean")
rank_courses(courses, grades, "median", descending=FALSE)
}
