% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frgtracker.R
\name{record_grades}
\alias{record_grades}
\title{Record grades for students of a specified course and its assessments.}
\usage{
record_grades(df)
}
\arguments{
\item{df}{A tidy data frame as a student gradebook,
with course id, student id, corresponding assessment id and grades.}
}
\value{
None
}
\description{
The grades are recorded to be out of 100.
}
