% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frgtracker.R
\name{suggest_grade_adjustment}
\alias{suggest_grade_adjustment}
\title{Adjust students' grades upwards in a course based on predefined benchmarks so
that average grades for those components meet or exceed these benchmarks.}
\usage{
suggest_grade_adjustment(
  courses,
  grades,
  id,
  benchmark_course = 90,
  benchmark_lab = 85,
  benchmark_quiz = 85
)
}
\arguments{
\item{courses}{A dataframe containing component weights for each course}

\item{grades}{A dataframe containing grades for students}

\item{id}{A string representing the course ID for which grades should
be adjusted.}

\item{benchmark_course}{A double value representing the benchmark of which
the average grade for the whole course must meet or exceed. Defaults to 90.}

\item{benchmark_lab}{A double value representing the benchmark of which the
average grade for each lab must meet or exceed. Defaults to 85.}

\item{benchmark_quiz}{A double value representing the benchmark of which the
average grade for each quiz must meet or exceed. Defaults to 85.}
}
\value{
A dataframe containing adjusted grades for all students in a course.
}
\description{
Students' grades in labs and quizzes will first be adjusted upwards to meet
or exceed benchmarks for labs and quizzes. However, if the average grade for
the course is still below the course benchmark, we will continue to adjust
upwards each lab / quiz until this benchmark is reached.
}
\examples{
grades <- data.frame(
course_id = c("511"),
student_id = c("tom"),
lab1 = c(100),
lab2 = c(80)
)
courses <- data.frame(
course_id = c("511"),
lab1 = c(0.45),
lab2 = c(0.55)
)
suggest_grade_adjustment(courses, grades, id = "511")
}
