% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frgtracker.R
\name{generate_course_statistics}
\alias{generate_course_statistics}
\title{Calculate the summary statistics for specified courses including mean, median and quantiles.}
\usage{
generate_course_statistics(course_ids)
}
\arguments{
\item{course_ids}{A list including all course IDs for which the summary statistics are
calculated for.}
}
\value{
A dataframe containing the summary statistics for specified courses
}
\description{
Calculate the summary statistics for specified courses including mean, median and quantiles.
}
\examples{

}
