% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frgtracker.R
\name{rank_courses}
\alias{rank_courses}
\title{Calculate students' course grades to rank courses in ascending/descending order by a
specified method.}
\usage{
rank_courses(
  method = c("method", "median", "lst-quantile", "3rd-quantile"),
  descending = True
)
}
\arguments{
\item{method}{one of "method", "median", "lst-quantile", "3rd-quantile", defining
the method for calculating the course rankings.}

\item{descending}{A logical value to decide if the rank should be in descending or
ascending order. Default to True}
}
\value{
A dataframe containing the rank for specified courses
}
\description{
Calculate students' course grades to rank courses in ascending/descending order by a
specified method.
}
\examples{

}
