% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_students.R
\name{rank_students}
\alias{rank_students}
\title{Calculate the average grade for all students and ranks them.}
\usage{
rank_students(course_id = "all", n = 3, ascending = FALSE)
}
\arguments{
\item{ascending}{}
}
\value{

}
\description{
Calculate the average grade for all students and ranks them.
}
