% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frgtracker.R
\name{record_grades}
\alias{record_grades}
\title{Record grades for students of a specified course and its assessments.}
\usage{
record_grades(df)
}
\arguments{
\item{df}{A tidy data frame as a student gradebook,
with course id, student id, corresponding assessment id and grades.}
}
\value{
None
}
\description{
The grades are recorded to be out of 100.
}
\examples{
grade_df <- data.frame(course_id=rep(511, 6),
student_id=rep("Kiki", 6),
assessment_id = c('lab1', 'lab2', 'lab3', 'lab4', 'quiz1', 'quiz2'),
grade=c(rep(92.1, 3), rep(80.2, 3)))

record_grades(grade_df)
}
