% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frgtracker.R
\name{register_courses}
\alias{register_courses}
\title{Read and store the input data frame into a data frame of courses to be
registered.}
\usage{
register_courses(df)
}
\arguments{
\item{df}{A tidy data frame containing course information,
with course id, all of the available assessments and corresponding weights.}
}
\value{
None
}
\description{
For each of the course, the weights of all assessments should sum up to 1.
The weights of individual assessment should be between 0 and 1.
}
